refactor: organize rust bridge files by component
The Rust bridge source and header directories mixed runtime utilities,
filesystem adapters, and format implementations in a flat layout.
Group them into runtime, filesystem, Iceberg, Lance, Paimon, Vortex,
and Talon directories, keeping lib.rs at the source root and the
predicate parser alongside its Vortex caller.

Preserve existing Rust module names through explicit source paths and
update C++ includes and Talon CXX generation paths to match the new
layout. The change is limited to file placement and path references.
