+++ cpp/src/format/bridge/rust/include/iceberg/iceberg_bridge.h
