+++ cpp/src/format/bridge/rust/include/lance/lance_bridge.h
