+++ cpp/src/format/bridge/rust/src/filesystem/filesystem_object_store.rs
