+++ cpp/src/format/bridge/rust/src/iceberg/iceberg_bridgeimpl.rs
