+++ cpp/src/format/bridge/rust/src/lance/lance_bridgeimpl.rs
