+++ cpp/src/format/bridge/rust/src/lance/lance_memory_estimator.rs
