+++ cpp/src/format/bridge/rust/src/lance/lance_object_store.rs
