+++ cpp/src/format/bridge/rust/src/paimon/paimon_bridgeimpl.rs
