+++ cpp/src/format/bridge/rust/src/runtime/bridge_util.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "bridge_util.h"
+#include "runtime/bridge_util.h"
 
 #include <cerrno>
 #include <charconv>
