+++ cpp/src/format/bridge/rust/src/runtime/rust_runtime.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "rust_runtime.h"
+#include "runtime/rust_runtime.h"
 
 #include "rust/cxx.h"
 #include "rust-bridge/lib.h"
