+++ cpp/src/format/bridge/rust/src/talon/talon_bridge.rs
