+++ cpp/src/format/bridge/rust/src/vortex/predicate_parser.rs
