+++ cpp/src/format/bridge/rust/src/vortex/vortex_bridgeimpl.rs
