+++ cpp/src/format/bridge/rust/src/vortex/vortex_layout_strategy_v2.rs
