+++ cpp/src/format/iceberg/iceberg_format.cpp
@@ -17,7 +17,7 @@
 #include "milvus-storage/format/iceberg/iceberg_format_reader.h"
 #include "milvus-storage/format/iceberg/iceberg_common.h"
 #include "milvus-storage/filesystem/fs.h"
-#include "iceberg_bridge.h"
+#include "iceberg/iceberg_bridge.h"
 
 namespace milvus_storage {
