+++ cpp/src/format/paimon/paimon_format_reader.cpp
@@ -31,7 +31,7 @@
 #include "milvus-storage/common/log.h"
 #include "milvus-storage/filesystem/fs.h"
 #include "milvus-storage/format/paimon/paimon_common.h"
-#include "paimon_bridge.h"
+#include "paimon/paimon_bridge.h"
 
 namespace milvus_storage::paimon {
 namespace {
