+++ cpp/test/format/external_table_test.cpp
@@ -47,8 +47,8 @@
 #include "milvus-storage/format/iceberg/iceberg_common.h"
 #include "milvus-storage/manifest.h"
 #include "milvus-storage/transaction/transaction.h"
-#include "lance_bridge.h"
-#include "iceberg_bridge.h"
+#include "lance/lance_bridge.h"
+#include "iceberg/iceberg_bridge.h"
 #include "test_env.h"
 
 namespace milvus_storage {
