+++ cpp/test/format/iceberg/iceberg_bridge_test.cpp
@@ -36,8 +36,8 @@
 #include "milvus-storage/filesystem/fs.h"
 #include "milvus-storage/filesystem/ffi/filesystem_internal.h"
 #include "milvus-storage/format/iceberg/iceberg_common.h"
-#include "bridge_util.h"
-#include "iceberg_bridge.h"
+#include "runtime/bridge_util.h"
+#include "iceberg/iceberg_bridge.h"
 #include "rust/cxx.h"
 #include "rust-bridge/lib.h"
 #include "test_env.h"
